requirements.txt
google-generativeai
PyPDF2
python-docx
pillow
pytesseract
opencv-python
